extract table row with error boundary to new component
